logging implementation
by slf4j logging ,we can know in console errror like which class or method gives an error when we use use that method on postman,the console shows which method gives and error or which class,for that we need to log(...) code in method or import slf4j. also we can print info trace warn these are level of log.